[rpc] Refactor tracer to a public API 	* Move tracing functions to hmy object 	* Add new "debug" namespace for tracing functions 	* Some tracing functions are disabled & return a RPC not available yet error
